Fix generation of file hashes for production
Building for production was producing files with different hashes
depending on the absolute path of the project. This was breaking
deployment because the static files are built once for the static
builder to serve them, and once for the node builder to retrieve the url
of the files to serve, but both builds were generating different file
names.