Fix runtime error
The project was being built on one machine with code that tried to read
the asset manifest at an absolute path on that machine, but was then run
on another machine where the path did not exist, resulting in a runtime
error.
This used to work because @now/node was compiling the project with ncc
which bundled everything, but it doesn't seem to do that anymore.
This fixes it by instead bundling the manifest with webpack.